requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
yfinance>=0.2.36
curl_cffi>=0.5.0
